Add migration to change wallet amount field from int to float. Tested API now working correctly.

+++ Persistence/Migrations/20191026160649_AddWalletModel.cs
@@ -14,7 +14,7 @@ namespace Persistence.Migrations
                     Id = table.Column<Guid>(nullable: false),
                     Type = table.Column<string>(nullable: true),
                     TransactionDate = table.Column<DateTime>(nullable: false),
-                    Amount = table.Column<int>(nullable: false),
+                    Amount = table.Column<float>(nullable: false),
                     AppUserId = table.Column<string>(nullable: false)
                 },
                 constraints: table =>
